pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dubbsystem"
version = "0.1.0"
description = "GPU-oriented open source video dubbing pipeline with transcription, translation, voice cloning, and muxing."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "DubbSystem" }
]
dependencies = [
  "ffmpeg-python>=0.2.0",
  "faster-whisper>=1.1.1",
  "numpy>=1.26.4",
  "pydantic>=2.7.0",
  "pydub>=0.25.1",
  "soundfile>=0.12.1",
  "librosa>=0.10.2.post1",
  "torch>=2.3.0",
  "torchaudio>=2.3.0",
  "transformers>=4.44.0",
  "sentencepiece>=0.2.0",
  "sacremoses>=0.1.1",
  "TTS>=0.22.0",
  "typer>=0.12.3",
  "rich>=13.7.1"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-mock>=3.14.0"
]

[project.scripts]
dubb = "dubb.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]